utils: add ast helpers
