+++ lib/rules/no-element-handle.js
@@ -1,29 +1,4 @@
-function isPageIdentifier({ callee }) {
-  return (
-    callee &&
-    callee.type === 'MemberExpression' &&
-    callee.object.type === 'Identifier' &&
-    callee.object.name === 'page'
-  );
-}
-
-function isElementHandleIdentifier({ callee }) {
-  return (
-    callee &&
-    callee.property &&
-    callee.property.type === 'Identifier' &&
-    callee.property.name === '$'
-  );
-}
-
-function isElementHandlesIdentifier({ callee }) {
-  return (
-    callee &&
-    callee.property &&
-    callee.property.type === 'Identifier' &&
-    callee.property.name === '$$'
-  );
-}
+const { isPageIdentifier, isElementHandleIdentifier, isElementHandlesIdentifier } = require("../utils/ast");
 
 function getRange(node) {
   const start = node.parent && node.parent.type === 'AwaitExpression'
