+++ lib/rules/no-page-pause.js
@@ -1,8 +1,10 @@
+const { isPageIdentifier, isPauseIdentifier } = require("../utils/ast");
+
 module.exports = {
   create(context) {
     return {
-      MemberExpression(node) {
-        if (node.object.name === "page" && node.property.name === "pause") {
+      CallExpression(node) {
+        if (isPageIdentifier(node) && isPauseIdentifier(node)) {
           context.report({ messageId: "noPagePause", node });
         }
       },
